[Design] added design to tips datatable
