+++ src/services/design.service.js
@@ -1,0 +1,23 @@
+const getStatusColor = (row) => {
+    if (row.tipStatus === 1) {
+      return "success"
+    }
+    else if (row.tipStatus === 0) {
+      return "danger"
+    }
+  }
+
+  const getStatusButtonText = (row) => {
+    if (row.tipStatus === 1) {
+      return "Enabled"
+    }
+    else if (row.tipStatus === 0) {
+      return "Disabled"
+    }
+  }
+  
+
+export {
+    getStatusButtonText,
+    getStatusColor
+  }
