+++ src/services/tips.service.js
@@ -1,6 +1,6 @@
 import axios from '../tools/axios'
 
-const token = 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9sb2dpbiIsImlhdCI6MTU5ODI4MzY1NSwiZXhwIjoxNTk4Mjg3MjU1LCJuYmYiOjE1OTgyODM2NTUsImp0aSI6InB3TzhrdmxUVTFUSU8wc3kiLCJzdWIiOjEsInBydiI6IjFiMmI3MzE0NzFhMjA0ZTRkYWY1MTRjYmJlYzNjZjhhNzk4OGU0YjMifQ.L-GjnUK3DFF_E2fh_GgQEdJFV7OBEIkpMPeJD4pyNw0'
+const token = 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6ODAwMFwvYXBpXC9sb2dpbiIsImlhdCI6MTU5ODQ2MTg4MSwiZXhwIjoxNTk4NDY1NDgxLCJuYmYiOjE1OTg0NjE4ODEsImp0aSI6ImxTT21IdGVTQ2t2UTRKcFAiLCJzdWIiOjEsInBydiI6IjFiMmI3MzE0NzFhMjA0ZTRkYWY1MTRjYmJlYzNjZjhhNzk4OGU0YjMifQ.gHrCLSVSGLbO7WdNN7EpCMLT64XDP2NZCucuVF5ObbY'
 const config = {
   headers: { Authorization: `Bearer ${token}` }
 };
